Dotfiles version

+++ kubox-init.sh
@@ -1,15 +1,21 @@
-#!/bin/sh
+#!/bin/bash
 
-set -ex
+DOTFILES_VERSION=v0.0.1
+DOTFILES_INIT=0
 
-mkdir -p ~/Repos
-cd ~/Repos
+if [[ ! -z ~/.dotfiles || cd ~/.dotfiles && $(git describe --tags)!=$DOTFILES_VERSION ]]; then
+    echo "Initializing kubox"
+    mkdir -p ~/Repos
+    cd ~/Repos
 
-rm -rf dotfiles
-git clone https://github.com/zzh8829/dotfiles.git
+    rm -rf dotfiles
+    git clone https://github.com/zzh8829/dotfiles.git
+    git checkout tags/v0.0.1
 
-cd dotfiles
-./linux.sh
-./dotfiles.sh
+    cd dotfiles
+    ./linux.sh
+    ./dotfiles.sh
+    ./vim.sh
+fi
 
-exec zsh -l
+sleep infinity
